Added PaymentMethod

+++ src/main/java/com/example/parkinglot/repo/PriceRepository.java
@@ -8,6 +8,6 @@ import java.util.List;
 
 @Repository
 public interface PriceRepository
-        extends CrudRepository<Price, Double>{
+        extends CrudRepository<Price, Integer>{
     List<Price> findAll();
 }
